requirements.txt
django==1.10.4
django-filter